refactor: move story utils to lib
